requirements.txt
#luigi
#luigi>=2.8.13
#pyyaml>=5.3.1
pyyaml
path.py
lsc@https://github.com/shouldsee/lsc/tarball/0.0.1
pytest
future-fstrings[rewrite]
attrdict
jinja2
graphviz